netplan: split off the bindings into their own module
The utils module was getting kind of crowded.
